re-using MLP inputs and bug fixes
